Multiple patches from jcleaver: - network tests resulting in odd connect()-errors now include the test definition in the error message - r6905 from trunk: fix RRD update of xymongen status (blanks must be removed) - sendmsg() debug code had syntax error - stackio: Must initialise eolchar - xymond: Timers for statistics update and purple-check changed from hardcoded values. - Clients now use "xymon --version" as the default for client-version data - xymoncmd: Use $HOSTNAME as the hostname, if defined, before running "uname". Strip newline from uname output. - xymoncmd: Fix buffer overflow in defining SERVEROSTYPE environment variable. - xymonnet and xymongen can now run via CRONDATE definitions. - xymonlaunch: Search for tasks.cfg in /etc, /etc/xymon, /etc/xymon-client and $XYMONHOME/etc - xymond_client: Permit data from unknown client types, but parse only the "messages" (logs), "file" and "linecount" data. xymond_client needs "--unknownclientosok" to enable this. - Linux client: Send error-output from netstat/ifconfig to /dev/null - loadhosts: Handle "vpage", "vsubpage" and "vsubparent" correctly. - loadhosts: Increase default hosts.cfg filebuffer size to 20 kB - xymond: Send "drop" and other admin-commands to the client-change and user-channels. - xymond_channel: Make the timeout for a worker to handle a message configurable (was 30 seconds hardcoded). - xymond_channel: Don't commit genocide on xymonlaunch when shutting down in network mode. - xymond_rrd: Add --no-rrd option to let all graph data flow to an external processor only. - Linux client: Format ps output to align columns better. - Add SHELL variable to xymonserver.cfg, and use it when we need to run an external command.
git-svn-id: svn://svn.code.sf.net/p/xymon/code@7204 44351d6e-118b-4698-b696-ce33095ecaa4

+++ branches/4.3.12/xymonnet/xymon-snmpcollect.c
@@ -26,7 +26,7 @@ static char rcsid[] = "$Id$";
 typedef struct oid_t {
 	mibdef_t *mib;				/* pointer to the mib definition for mibs */
 	oid Oid[MAX_OID_LEN];			/* the internal OID representation */
-	unsigned int OidLen;			/* size of the oid */
+	size_t OidLen;			/* size of the oid */
 	char *devname;				/* Users' chosen device name. May be a key (e.g "eth0") */
 	oidds_t *oiddef;			/* Points to the oidds_t definition */
 	int  setnumber;				/* All vars fetched in one PDU's have the same setnumber */
@@ -126,7 +126,7 @@ struct snmp_pdu *generate_datarequest(req_t *item)
 int print_result (int status, req_t *req, struct snmp_pdu *pdu)
 {
 	struct variable_list *vp;
-	int len;
+	size_t len;
 	keyrecord_t *kwalk;
 	int keyoidlen;
 	oid_t *owalk;
@@ -136,7 +136,7 @@ int print_result (int status, req_t *req, struct snmp_pdu *pdu)
 	  case STAT_SUCCESS:
 		if (pdu->errstat == SNMP_ERR_NOERROR) {
 			unsigned char *valstr = NULL, *oidstr = NULL;
-			int valsz = 0, oidsz = 0;
+			size_t valsz = 0, oidsz = 0;
 
 			okcount++;
 
@@ -545,7 +545,7 @@ void stophosts(void)
 void setupmib(mibdef_t *mib, int verbose)
 {
 	mibidx_t *iwalk;
-	int sz, len;
+	size_t sz, len;
 
 	if (mib->loadstatus != MIB_STATUS_NOTLOADED) return;
